console: Move api to separate folder
