+++ pkg/webui/console/api/stubs/fake-data.js
