+++ pkg/webui/console/api/stubs/index.js
