+++ pkg/webui/console/store/actions/application.js
