+++ pkg/webui/console/store/actions/applications.js
