+++ pkg/webui/console/store/actions/devices.js
