+++ pkg/webui/console/store/actions/rights.js
