+++ pkg/webui/console/store/index.js
@@ -18,8 +18,8 @@ import { createStore, applyMiddleware, compose } from 'redux'
 import { createLogicMiddleware } from 'redux-logic'
 import { connectRouter, routerMiddleware } from 'connected-react-router'
 
-import reducer from '../reducers'
-import logic from '../logic'
+import reducer from './reducers'
+import logic from './middleware'
 
 const composeEnhancers = (process.env.NODE_ENV === 'development'
   && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose
