+++ pkg/webui/console/store/reducers/application.js
