+++ pkg/webui/console/store/reducers/applications.js
