+++ pkg/webui/console/store/reducers/client.js
