+++ pkg/webui/console/store/reducers/devices.js
