+++ pkg/webui/console/store/reducers/gateways.js
