+++ pkg/webui/console/store/reducers/rights.js
