+++ pkg/webui/console/views/application-access-list/index.js
@@ -18,7 +18,7 @@ import bind from 'autobind-decorator'
 
 import IntlHelmet from '../../../lib/components/intl-helmet'
 import ApiKeysTable from '../../../containers/api-keys-table'
-import { getApplicationApiKeysList } from '../../../actions/application'
+import { getApplicationApiKeysList } from '../../store/actions/application'
 import sharedMessages from '../../../lib/shared-messages'
 
 const API_KEYS_TABLE_SIZE = 10
