+++ pkg/webui/containers/devices-table/index.js
@@ -21,7 +21,7 @@ import sharedMessages from '../../lib/shared-messages'
 import Message from '../../lib/components/message'
 import FetchTable from '../fetch-table'
 
-import { getDevicesList, searchDevicesList } from '../../actions/devices'
+import { getDevicesList, searchDevicesList } from '../../console/store/actions/devices'
 
 const m = defineMessages({
   deviceId: 'Device ID',
